--- src/components/Error/Error.ts
@@ -1,9 +1,0 @@
-import renderHandler from './render';
-
-function Error(errorStatus: number) {
-  const errorComponent = renderHandler(errorStatus);
-
-  return errorComponent;
-}
-
-export default Error;
