+++ src/components/ErrorComponent/ErrorComponent.ts
@@ -1,0 +1,10 @@
+import renderHandler from './render';
+
+function ErrorComponent(errorStatus: number) {
+  console.log('ErrorComponent :', errorStatus);
+  const errorComponent = renderHandler(errorStatus);
+
+  return errorComponent;
+}
+
+export default ErrorComponent;
